backend/requirements.txt
Django>=5.0
djangorestframework
django-unfold
django-cors-headers
psycopg2-binary
pillow
replicate
